Rest package polish (#539)
* Split generic rest-functionality to utils, add test

* Handle empty bodies

* Do not set mimetype for empty request bodies
